refactor(FR-2015): restructure E2E test directory organization and standardize tagging (#5195)
Resolves #5213 ([FR-2015](https://lablup.atlassian.net/browse/FR-2015))

## Summary

Reorganize E2E test files into feature-based directories and implement consistent test tagging strategy.

### Directory Restructuring
- Created feature-based directories: `auth/`, `user/`, `vfolder/`, `session/`, `serving/`, `environment/`, `agent/`, `maintenance/`, `app-launcher/`, `config/`
- Moved 15 test files to appropriate feature directories
- Renamed `.test.ts` files to `.spec.ts` for consistency
- Moved 8 POM classes to feature-specific subdirectories (`session/`, `user/`, `vfolder/`, `common/`)
- Removed empty `critical/` directories (replaced by tags)

### Consistent Tagging
- Added priority tags: `@smoke` (PR checks), `@critical` (pre-merge), `@regression` (nightly)
- Added feature tags: `@auth`, `@user`, `@vfolder`, `@session`, `@serving`, `@environment`, `@agent`, `@maintenance`, `@app-launcher`, `@config`
- Added type tag: `@functional`

### Documentation
- Updated `e2e/README.md` to reflect new structure
- Added file naming conventions and contributing guidelines

### Files Changed
- 45 files changed, 5831 insertions(+), 2979 deletions(-)

## Test plan

- [ ] Verify tests can be listed by tag: `pnpm exec playwright test --list --grep @smoke`
- [ ] Verify tests can be run by feature directory: `pnpm exec playwright test e2e/auth/`
- [ ] Verify ESLint passes for all E2E test files

[FR-2015]: https://lablup.atlassian.net/browse/FR-2015?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
