+++ e2e/utils/classes/session/AppLauncherModal.ts
