+++ e2e/utils/classes/vfolder/FolderExplorerModal.ts
